Updating the tests for the split themes
